refactor: move stop_by to separate file
